Clarify and add new demo
