+++ common/lib/common-definitions/src/logger.ts
@@ -120,23 +120,3 @@ export interface ITelemetryLogger extends ITelemetryBaseLogger {
      */
     shipAssert(condition: boolean, event?: ITelemetryErrorEvent): void;
 }
-
-// ////////////////////// fluidError.ts in common-definitions /////////////////////////////
-/**
- * Can be decl merged to add stuff like packagename, odspErrorResponse, etc
- * The logger impl can choose to log anything in here as appropriate, but by default it's not
- */
-export interface ISensitiveDebugData {
-    innerError?: any;
-}
-
-export interface IFluidError {
-    errorType: string;
-    message: string;
-    getFluidTelemetryProps: () => ITelemetryProperties;  //* use property getters?
-    getSensitiveDebugData: () => ISensitiveDebugData & { stack: string };
-    addDetails: (props: ITelemetryProperties, debugData: ISensitiveDebugData) => void;
-}
-
-export const isIFluidError = (err: any): err is IFluidError =>
-    typeof(err?.getFluidTelemetryProps) === "function";
