+++ packages/runtime/container-runtime/src/dataStoreContext.ts
@@ -67,6 +67,13 @@ import {
     DataStoreSnapshotFormatVersion,
 } from "./snapshot";
 
+//* Would probably go in container-runtime-definitions package
+declare module "@fluidframework/common-definitions" {
+    export interface ISensitiveDebugData {
+        packageName?: string;
+    }
+}
+
 function createAttributes(pkg: readonly string[], isRootDataStore: boolean): IFluidDataStoreAttributes {
     const stringifiedPkg = JSON.stringify(pkg);
     return {
@@ -262,11 +269,8 @@ export abstract class FluidDataStoreContext extends TypedEventEmitter<IFluidData
         }
     }
 
-    private rejectDeferredRealize(reason: string): never {
-        const error = new Error(reason);
-        // Error messages contain package names that is considered Personal Identifiable Information
-        // Mark it as such, so that if it ever reaches telemetry pipeline, it has a chance to remove it.
-        (error as any).containsPII = true;
+    private rejectDeferredRealize(reason: string, packageName?: string): never {
+        const error = new GenericFluidError(reason, {}, { packageName });
         throw error;
     }
 
@@ -289,18 +293,18 @@ export abstract class FluidDataStoreContext extends TypedEventEmitter<IFluidData
         let lastPkg: string | undefined;
         for (const pkg of packages) {
             if (!registry) {
-                this.rejectDeferredRealize(`No registry for ${lastPkg} package`);
+                this.rejectDeferredRealize(`No registry for package`, lastPkg);
             }
             lastPkg = pkg;
             entry = await registry.get(pkg);
             if (!entry) {
-                this.rejectDeferredRealize(`Registry does not contain entry for the package ${pkg}`);
+                this.rejectDeferredRealize(`Registry does not contain entry for the package`, pkg);
             }
             registry = entry.IFluidDataStoreRegistry;
         }
         const factory = entry?.IFluidDataStoreFactory;
         if (factory === undefined) {
-            this.rejectDeferredRealize(`Can't find factory for ${lastPkg} package`);
+            this.rejectDeferredRealize(`Can't find factory for package`, lastPkg);
         }
 
         return { factory, registry };
